refactor: review updates
