+++ packages/scaffold-ui/src/partials/w3m-connector-list/index.ts
@@ -45,7 +45,7 @@ export class W3mConnectorList extends LitElement {
   public override render() {
     const { custom, recent, announced, injected, multiChain, recommended, featured, external } =
       ConnectorUtil.getConnectorsByType(this.connectors)
-    const isConnectedWithWC = ConnectorController.getIsConnectedWithWC()
+    const isConnectedWithWC = ConnectorUtil.getIsConnectedWithWC()
     const enableWalletConnect = OptionsController.state.enableWalletConnect
 
     return html`
