+++ packages/scaffold-ui/src/utils/ConnectorUtil.ts
@@ -1,5 +1,7 @@
+import { ConstantsUtil } from '@reown/appkit-common'
 import {
   ApiController,
+  ChainController,
   ConnectionController,
   type ConnectorWithProviders,
   CoreHelperUtil,
@@ -35,6 +37,7 @@ export const ConnectorUtil = {
       featured: filteredFeatured
     }
   },
+
   showConnector(connector: ConnectorWithProviders) {
     if (connector.type === 'INJECTED') {
       if (!CoreHelperUtil.isMobile() && connector.name === 'Browser Wallet') {
@@ -63,5 +66,20 @@ export const ConnectorUtil = {
     }
 
     return true
+  },
+
+  /**
+   * Returns true if the user is connected to a WalletConnect connector in the any of the available namespaces.
+   * @returns boolean
+   */
+  getIsConnectedWithWC() {
+    const chains = Array.from(ChainController.state.chains.values())
+    const isConnectedWithWC = chains.some(chain => {
+      const connectorId = StorageUtil.getConnectedConnectorId(chain.namespace)
+
+      return connectorId === ConstantsUtil.CONNECTOR_ID.WALLET_CONNECT
+    })
+
+    return isConnectedWithWC
   }
 }
